refactor(ui): move Settings from block to component
Relocates the Settings compound component out of `mosaic/block` into `mosaic/components/settings`, and moves its swingset docs entry from the Blocks layer to Components.
