+++ packages/swingset/src/stories/settings.mdx
@@ -2,12 +2,12 @@ import * as SettingsStories from './settings.stories';
 
 # Settings
 
-A settings block that carries the canonical section, surface, row grouping, and item layout. Use it for product settings while composing each item's media, content, and actions as needed.
+A settings component that carries the canonical section, surface, row grouping, and item layout. Use it for product settings while composing each item's media, content, and actions as needed.
 
 ## Usage
 
 ```tsx
-import { Settings } from '@clerk/ui/mosaic/block/settings';
+import { Settings } from '@clerk/ui/mosaic/components/settings';
 
 <Settings.Root>
   <Settings.Title>Account</Settings.Title>
