+++ packages/ui/src/mosaic/components/settings/index.ts
@@ -1,0 +1,15 @@
+export { Settings, settingsVars } from './settings';
+export type {
+  SettingsActionsProps,
+  SettingsContentProps,
+  SettingsDescriptionProps,
+  SettingsGroupProps,
+  SettingsItemProps,
+  SettingsItemsProps,
+  SettingsLabelProps,
+  SettingsMediaProps,
+  SettingsMediaSize,
+  SettingsRootProps,
+  SettingsRowProps,
+  SettingsTitleProps,
+} from './settings';
