+++ packages/ui/src/mosaic/components/settings/settings.vars.stylex.ts
@@ -1,6 +1,6 @@
 import * as stylex from '@stylexjs/stylex';
 
-import { colorVars, space } from '../tokens.stylex';
+import { colorVars, space } from '../../tokens.stylex';
 
 export const settingsVars = stylex.defineVars({
   '--cl-settings-background': colorVars['--cl-color-background'],
